[Xamarin.Android.Build.Tasks] Emit @(XamarinProjectAnalysisResult) (#1355)
Visual Studio allows "project analysis rules" to be run against a
project, so that problems which shouldn't be build errors or warnigns
can still be reported to the developer.

In Visual Studio for Mac, the analysis rules can be executed via the
**Project** > **Run Code Analysis On *Project Name*** menu item.

However, when I do this in a project which should elicit an analysis
error, *nothing* is reported. :-/

The `xamarin-analysis` documentation states that the analysis rules
should emit errors into the `@(XamarinProjectAnalysisResult)`
item group.

We were inadvertently placing these into
`@(XamarinPackageAnalysisTargets)`, which is actually used to
determine which targets should be executed. As such, no errors are
actually displayed to the developer within the IDE.