Fix package creation on oss
`make package-oss` fails on Linux with the following error:

    if [ ! -f xamarin.android-oss_v8.3.99.23_Linux-x86_64_master_3f587f51/ThirdPartyNotices.txt ] ; then cp -f bin/*/lib/xamarin.android/ThirdPartyNotices.txt xamarin.android-oss_v8.3.99.23_Linux-x86_64_master_3f587f51 ; fi
    cp: will not overwrite just-created 'xamarin.android-oss_v8.3.99.23_Linux-x86_64_master_3f587f51/ThirdPartyNotices.txt' with 'bin/Release/lib/xamarin.android/ThirdPartyNotices.txt'

which is due to some weird decisions GNU cp makes when copying files with the
same names from different source directories to the same destination. The fix is
to use `-a` (archive), however, it's not portable as macOS (BSD) version of cp
doesn't support it. Thus, this patch