[Xamarin.Android.Build.Tasks] The "ConvertResourcesCases" task failed unexpectedly.
DevDiv Issue 571365

We have a report that `ConvertResourcesCases` is failing with the follow
error.

	error MSB4018: The "ConvertResourcesCases" task failed unexpectedly.
	Android/Xamarin.Android.Common.targets(1334,2): error MSB4018: System.IO.FileNotFoundException: <some path>/design_layout_snackbar.xml.tmp does not exist

This is completely weird, especially as the Task is not called in
parallel. What I susepct is happening is VSForMac is running is
UpdateResources task at the same time the user is doing a build.

Because the name of the temp files we are just `foo.xml.tmp` it
is highly possible that one instance of the task is deleting the
file while the other instance is still running. This is difficult
to replicate.

This change uses

	System.IO.Path.GetFileName (System.IO.Path.GetTempFileName ());

to generate a random temp file name for the temp file. This should
stop filename collisions.