[build] Fix package creation on oss (#1363)
`make package-oss` [fails on Linux][0] with the following error:

[0]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android-linux/820/

    if [ ! -f xamarin.android-oss_v8.3.99.23_Linux-x86_64_master_3f587f51/ThirdPartyNotices.txt ] ; then cp -f bin/*/lib/xamarin.android/ThirdPartyNotices.txt xamarin.android-oss_v8.3.99.23_Linux-x86_64_master_3f587f51 ; fi
    cp: will not overwrite just-created 'xamarin.android-oss_v8.3.99.23_Linux-x86_64_master_3f587f51/ThirdPartyNotices.txt' with 'bin/Release/lib/xamarin.android/ThirdPartyNotices.txt'

which is due to some weird decisions GNU cp makes when copying files
with the same names from different source directories to the same
destination. The fix is to use `cp -a` (archive). However, `cp -a` is
not portable as macOS (BSD) version of cp doesn't support it.

Use `cp -f` or `cp -af` depending on the host `$(OS_NAME)`.