[build] Add `make package-test-errors` target (#1366)
Sometimes a PR build will [fail][0]. That's to be expected! ;-)

[0]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android-pr-builder/2600/

However, sometimes the error message is [largely useless][1]:

[1]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android-pr-builder/2600/testReport/junit/Xamarin.Android.Build.Tests.BuildTest/BuildAotApplicationAndBundle/Xamarin_Android_Build_Tests_BuildTest_BuildAotApplicationAndBundle__armeabi_v7a__False_True____Debug/

	MESSAGE:
	  Build should have succeeded.
	Expected: True
	But was:  False

	STACK TRACE:
	  at Xamarin.Android.Build.Tests.BuildTest.BuildAotApplicationAndBundle (System.String supportedAbis, System.Boolean enableLLVM, System.Boolean expectedResult)

That's not helpful at all.

In such circumstanes, the usual "fallback" path is to check the
Jenkins [**Workspace**][2], then pray that the mentioned failing
projects *still exist*. The problem here is that the Workspace is
cleared on *every* Jenkins build (for a specific job), so if another
PR build has started, the Workspace for the previous build is no
longer accessible.

[2]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android-pr-builder/ws/

Add a new `make package-test-errors` target. When called, *if* there
are any error projects within `bin/Test*/temp`, they will be packaged
up into `test-errors.zip`. We can then upload `test-errors.zip` to
Azure -- along with the other PR build outputs -- allowing for saner
determination of What Went Wrong™, allowing the PR to be fixed.