[Xamarin.Android.Build.Tasks] Add @(AndroidAarLibrary)
Adding `@(AndroidAarLibrary)` to `@(AvailableItemName)` *should* have
been done as part of commit 6a8ea2bb, but was overlooked.

We want `@(AvailableItemName)` to be up-to-date so that you can have
conditional MSBuild behavior based on whether a particular item is
supported.

In particular, @Redth wants to ensure that the `@(AndroidAarLibrary)`
build action is available as part of his Components work, and ensuring
that `@(AvailableItemName)` contains `@(AndroidAarLibrary)` is the
most straightforward way to do so.

Additionally, sort and re-indent the `@(AvailableItemName)` entries.