[Xamarin.Android.Build.Tasks] Fix an issue with `actionLayout`
Our Resource Case converter `AndroidResource` did not handle
the `actionLayout` attribute for the `http://schemas.android.com/apk/res-auto`
namespace in menu xml files.

As a result the value for the `actionLayout` would not be
lowercased. This results in the following error

	No resource found that matches the given name (at ‘actionLayout’ with value ‘@layout/ServingLayout’).

The fix is to add support for `actionLayout` to `AndroidResource`.
We also have added a unit test for this particular case and allowed
the `Xamarin.Android.Build.Tests` access to the internals of the
`Xamarin.Android.Build.Tasks` so it can unit test more of the
internal classes it has.