[Xamarin.Android.Build.Tasks] The "ConvertResourcesCases" task failed unexpectedly. (#1371)
Fixes? https://devdiv.visualstudio.com/DevDiv/_workitems/edit/571365

We have a report that `ConvertResourcesCases` is failing with the
following error:

	error MSB4018: The "ConvertResourcesCases" task failed unexpectedly.
	Android/Xamarin.Android.Common.targets(1334,2): error MSB4018: System.IO.FileNotFoundException: <some path>/design_layout_snackbar.xml.tmp does not exist

This is completely weird, especially as the Task is not being called
concurrently (or so we hope & assume). What I suspect is happening is
Visual Studio for Mac is running the `UpdateAndroidResources` target
at the same time the user started a build.

Because the name of the temp files was just `foo.xml.tmp` it is
highly possible that one instance of the task is deleting the file
while another instance is still running. This is difficult to
replicate.

Attempt to "fix" things by using `System.IO.Path.GetTempFileName()`
instead of appending `.tmp` to the filename. This should avoid
filename collisions.

However, if our suspicion is correct that the `UpdateAndroidResources`
target is concurrently executing alongside a Build, this patch will
*not* fix the concurrent target execution scenario.