[Xamarin.Android.Build.Tasks] CodeBehind fragment & widget access (#1378)
Fragment handling code partially taken from:

	https://github.com/xamarin/xamarin-android/pull/1302/

Fragments are treated as normal widget, with the difference that they
have to be found using `FragmentManager.FindFragmentById()` instead
of the usual `Activity.FindViewById()` method.

Each generated class also gets a property named `Widget` which
returns the actual Android widget as found in the layout file. This
allows us to keep hierarchical nature of the XML code (thus handling
nested widgets with duplicate ids gracefully) while being able to
access the parent widget itself.

The commit introduces a new attribute called `tools:managedType`
which is used to specify the element associated property's type. We
cannot use `tools:class` for this purpose since the layout root
element already uses it to specify the name of the generated
code-behind class and if the element had `android:id` on it, we would
end up using the activity's type for the root element's property in
the generated code, instead of its actual type.