[Xamarin.Android.Build.Tasks] Skip InstallationID when unavailable (#1388)
This is required to support the new "Visual Studio Build Tools" SKU
(installs only build tools) where `devenv.isolation.ini` is not
available (already confirmed this is intentional with the VS Setup
team) and thus our build breaks. We are using InstallationID to
support SxS at design-time so this isn't really needed in this case,
it's safe to skip it.