Use `cp -n` during ThirdPartyNotices.txt copying.
`cp -f` (or `-af`) doesn't help on Linux when copying multiple files with
the same name to one target directory (as is the case here when building
both Debug and Release configurations). Using `cp -n` means the overwrite
is skipped, rather than fatal. macOS does not care about the overwrite
either way - passing `-n` to BSD `cp` is non-fatal, and fixes the build
failures on Linux.