Bump Android Emulator to version 27.1.12
This is the latest emulator from Google. Faster and more reliable.
Unfortunately, on Ubuntu it is necessary to remove the bundled copy
of the libstd++ library since it conflicts with the system-installed
version and prevents the emulator from starting.

The removal is limited to Ubuntu because that's the only distribution
we tested it on. Technically it is needed only on Ubuntu 17.10 and
newer since the emulator works fine on the current LTS 16.04, but since
removing it on 16.04 doesn't hurt it in any way we can skip checking the
exact version of the OS and make the condition simpler.