[build-tools] Improve logcat timing processing
Add another regexp match for timing log start. So far I have been
using the logcat timing with logcat output from the emulator with API
version 21. Looks like the logcat format changed in newer Android
versions.

Here is an example of the relevant part of the logcat output on Pixel
2 device (API 27):

```
03-08 15:11:48.140  1128  4745 I ActivityManager: Start proc 18503:com.xamarin.xatemplateaot/u0a241 for activity com.xamarin.xatemplateaot/md58018c7d08c228c8e2a3e03723c59ca27.MainActivity
```