[xabuild] Ensure directory removal with rm `-rf`. (#1407)
This is the common build failure, what we see after `make prepare` and `make`:

	rm: cannot remove '/sources/monodroid/monodroid/external/xamarin-android/tools/scripts/../../bin/Debug/lib/xamarin.android/xbuild/Xamarin': Is a directory

This does not show up once you run `make` again, so it's a cosmetic problem.
But this should not happen. First-time contributors would get stuck at it
and may leave after seeing this. This should be fixed.