[Xamarin.Android.Build.Tasks] Fix <Proguard /> w/ acw-map.txt (#1402)
Fixes: https://github.com/xamarin/xamarin-android/issues/1373

Since e5b1c92c, the contents of the `acw-map.txt` file have changed. We
have removed the fully qualified type names from the file, and there is
some code in the `<Proguard />` task that relies on a specific number of
lines per C# type.

We need to update the for-loop to skip every third line instead of every
fourth line.

I also updated the loop to only index against the `acwLines` array once
per loop interation, as a small performance improvement. The code is also
a little more readable from this change.

Lastly, I added some code to the proguard test so that it validates the
proguard configuration file contains the appropriate contents. Prior to
this, the test was only making sure the build succeeded.