[build] Windows support for ThirdPartyNotices
Context: #464

Since the `ThirdPartyNotices.txt` functionality was implemented in
`Makefile`, this file was not getting generated on Windows. This commit
adds the same functionality to `PrepareWindows.targets`, which basically
just builds `ThirdPartyNotices.csproj` twice with different parameters.

The only workaround I had to add for Windows was to make the
`<ProjectReference />` for `xa-prep-tasks.csproj` condition on Windows.
Otherwise, I was getting into a file-locking issue where the nested
`<MSBuild />` task was trying to copy `xa-prep-tasks.dll` to the output
directory while it was in use.