[build-tools] Improve logcat timing processing (#1397)
Add another regular expression match for the timing log start. So far
I have been using the logcat timing with logcat output from the
emulator with API version 21. It looks like the logcat output from
`ActivityManager` has changed in newer Android versions.

Here is an example of the relevant part of the logcat output on a
Pixel 2 device (API-27):

	03-08 15:11:48.140  1128  4745 I ActivityManager: Start proc 18503:com.xamarin.xatemplateaot/u0a241 for activity com.xamarin.xatemplateaot/md58018c7d08c228c8e2a3e03723c59ca27.MainActivity

While on the emulator with API 21 we see:

	03-07 20:53:48.816  1527  1869 I ActivityManager: Start proc com.xamarin.xatemplateaot for activity com.xamarin.xatemplateaot/md58018c7d08c228c8e2a3e03723c59ca27.MainActivity: pid=3080 uid=10057 gids={50057, 9997, 3003} abi=x86

The main difference here for us is the location of the PID number.