Discard POST content on request redirection
Fixes: https://github.com/xamarin/xamarin-android/issues/1282

When the original request uses the POST method and has a content payload, we
must discard the content when the server returns a redirection status as the
method changes to GET and the POST content cannot be used in this context.

This is not only the correct client behavior but it also prevents attempt to use
disposed content stream on the redirected request.