[msbuild] fix for JavaSourceJar and test
Context #1091

The test `BindingBuildTest.JavaSourceJar` appears to pass under xbuild,
but fail under msbuild. This seems be true for macOS or Windows.

Under further inspection, I noticed the `BuildDocumentation` target was
never running under msbuild. It was being skipped due to the outputs
being up-to-date in comparison to the inputs. To fix the skipped target,
I was able to add an additional input in
`Xamarin.Android.Bindings.targets`.

The line:
    <Copy
      SourceFiles="@(IntermediateAssembly->'$(IntermediateOutputPath)%(filename).xml')"
      DestinationFiles="@(IntermediateAssembly->'$(OutputPath)%(filename).xml')"
      SkipUnchangedFiles="true" />

Implies that `@(IntermediateAssembly->'$(IntermediateOutputPath)%(filename).xml')`
also should be an input.

Additionally, there was some stuff in the test that seemed incorrect:
- The return value of `bindingBuilder.Build` should have an assertion
that it is true
- For some reason the text from `GetIntermediaryAsText` was being passed
to `Path.Combine`. I think this was a typo that wasn't intentional.