[msbuild] fix for JavaSourceJar and test (#1420)
Context: d46abe823a40f026ee0cafd44a9e4c0a538fba4e

The test `BindingBuildTest.JavaSourceJar` appears to pass under
`xbuild`, but fails under `msbuild`. This seems be true for macOS
or Windows.

Under further inspection, I noticed the `BuildDocumentation` target
was never running under `msbuild`: it was being skipped due to the
Outputs being up-to-date in comparison to the Inputs. To fix the
skipped target,I was able to add an additional input in
`Xamarin.Android.Bindings.targets`.

The line:

    <Copy
        SourceFiles="@(IntermediateAssembly->'$(IntermediateOutputPath)%(filename).xml')"
        DestinationFiles="@(IntermediateAssembly->'$(OutputPath)%(filename).xml')"
        SkipUnchangedFiles="true" />

Implies that this should also be an Input:

    `@(IntermediateAssembly->'$(IntermediateOutputPath)%(filename).xml')`

Additionally, there was some stuff in the test that seemed incorrect:

  * The return value of `bindingBuilder.Build` should have an
    assertion that it is true
  * For some reason the text from `GetIntermediaryAsText()` was being
    passed to `Path.Combine()`. I think this was a typo that wasn't
    intentional.