[create-vsix] Create a separate package for unstable frameworks
**Background**

It's Android preview time!

Even better, we have support for building API-P bindings within around
a week of the Google announcement, via 8ce25376.

This raises a question: How does this binding get *used*?

Historically, we haven't been very good at this: various bits of code
needed to know *all* of the API levels (to associate with
`$(TargetFrameworkVersion)` values and Ids and names and...), which
meant that the only way the preview binding could be used was to have
a preview release of the *entire development stack*: Xamarin.Android,
IDE, build-tools, the works.

Invariably we didn't have a "preview" release of the new binding until
after it was API-stable (or close to it), and the IDEs have been
updated, all of which would be *months* after the announcement.

Commit 8942eca0 envisioned a New World Order™ in which the mapping
between API level and `$(TargetFrameworkVersion)`/etc. could be more
*dynamic*, based on `AndroidApiInfo.xml` files probed at process and
IDE startup, to *simplify* what it takes to use binding assemblies.

This support, mutated in various forms, is already in the IDEs (though
a preview IDE may be needed).

With foundational support, we can now try the next bit: update
`create-vsix.csproj` so that it can generate one of two packages:

 1. A `Xamarin.Android.Sdk-OSS-*.vsix` file which contains *only* the
    stable API level bindings.

 2. An optional `Xamarin.Android.Sdk-OSS-*-UnstableFrameworks.vsix`
    file which contains *only* the ***unstable*** API bindings.

The idea is that the first package can be the "normal" release
package, while the second package can be installed "side-by-side" with
the first pacakge.

Furthermore, the *intent* is that the `-UnstableFrameworks` package
could be installed side-by-side with *any* Xamarin.Android.Sdk
package. Ideally, this means that it could be installed alongside the
current *stable* Xamarin.Android SDK package, with no ill effects.

The target workflow is thus:

 1. Developer has Xamarin.Android SDK installed on their machine.

 2. Developer downloads and installs the `-UnstableFrameworks`
    package from:

        https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/lastSuccessfulBuild/Azure/

 3. Developer restarts their IDE.

 4. Developer edits the `.csproj` of their project, and sets the
    `$(TargetFrameworkVersion)` value to the value of the unstable API
    binding, e.g. `v8.1.99`.

    *Note*: It may instead be possible to change things from within
    the IDE. Some specifics are not yet sorted out.

 5. The next build of the project uses the preview/unstable API level.

With luck, this should allow us to provide "usable" (if broken!
they're not API stable!) for testing on an improved timeframe.

**Implementation**

Hack up `create-vsix.csproj` and company to support a new MSBuild
property: `$(UnstableFrameworks)` (default False). When True, the
output package name is changed to have a `-UnstableFrameworks` suffix,
the package ID is changed, the installation directory is changed...
*Lots* of things are changed, including the package contents.

The "normal" package is altered so that it *only* contains Stable
framework bindings. (This is a change from before, in which *all*
framework bindings were included.)

The "unstable" package contains only non-Stable framework bindings.

The `make create-vsix` Makefile target and `HaveUnstableFrameworks`
target within `build-tools/scripts/Info.targets` work together to only
build the `-UnstableFrameworks.vsix` package if there *are* unstable
frameworks to package. If there are no unstable packages, no
`-UnstableFrameworks` package is created.