[tests] change to re-save project file if properties change
Context: https://devdiv.visualstudio.com/DevDiv/_build/index?buildId=1480184&tab=ms.vss-test-web.test-result-details&_a=summary

In the case of `Xamarin.Android.Build.Tests.BuildTest.ValidateUseLatestAndroid`,
there are points at which `UseLatestPlatformSdk` and `TargetFrameworkVersion`
are changed, but the `csproj` file is not re-saved on Windows.

So how does the test pass on macOS?

They pass on macOS due to `FileInfo.LastWriteTimeUtc` not containing
the millisecond information, which means files are likely to re-save
on macOS when not needed.

A simple fix for Windows, is to keep a reference to the last `csproj`
file contents. If the contents change, make the `Timestamp` of the
`csproj` null so it will always re-save.

After making this change, I was able to remove the following code from
another unit test:

    proj.Sources.First ().Timestamp = DateTimeOffset.UtcNow;