Fix loading of Xamarin.Android.sln w/ omnisharp-roslyn (#1425)
Omnisharp-Roslyn uses an internal MSBuild loader which has a number of
demands/assumptions on the contents of `.csproj`+`.targets` files.
It assumes the following targets are present:

  * Compile
  * CoreCompile

It also assumes that the `$(TargetFrameworkMoniker)` property is set.

This commit makes a number of changes to our non-managed projects to
make sure the above expectations are met. With those changes in place
one can load `Xamarin.Android.sln` into an editor that uses OmniSharp
for .NET code completion etc.