[Xamarin.Android.Build.Tasks] JavaMaximumHeapSize Default (#1436)
Fixes: https://github.com/xamarin/xamarin-android/issues/1415

Commonly Xamarin.Android apps on Windows require the
`$(JavaMaximumHeapSize)` property to be set, in order to make it past
the `CompileToDalvik` MSBuild task that invokes `dx.jar`. Passing
`java.exe` a value of `-Xmx1G` often avoids the
`java.lang.OutOfMemoryError`

We have been getting reports of this happening for
**File** > **New Project** templates, so it seems to be time to
default the value to `1G` if it is not set.

Changes:

  - Set `$(JavaMaximumHeapSize)` in `Xamarin.Android.Common.targets`
	  if not overridden.
  - Removed places `$(JavaMaximumHeapSize)` was required to be set
	  for some unit tests to pass on Windows
  - Update documentation

Concerns:

At one time, a default value for `$(JavaMaximumHeapSize)` was
attempted, only to get [reports of an error][0] such as:

	JVMJ9VM015W Initialization error for library j9gc24(2): Failed to instantiate heap; 1G requested
	Could not create the Java virtual machine.

Hopefully, this was something that was happening for older JDK
versions in the past on Windows.

[0]: https://www.ibm.com/developerworks/community/blogs/aimsupport/entry/why_wont_my_jvm_start_with_this_heap_size?lang=en