[Xamarin.Android.Build.Tasks] Rework Parallel.ForEach (#1434)
We are getting lots of reports of hangs in the `aapt` task. While
this might be down to the `aapt` version itself, it did highlight a
possible problem with our implementation.

Firstly `Parallel.ForEach()` is a ***BLOCKING*** call, and our
implementation in `Aapt.cs` and `Crunch.cs` was not taking that into
account. Also `Parellel.ForEach()` will use the current thread to run
tasks. If we run it from the main MSBuild `Execute()` method it will
run at least one of the items on that thread. Given one of our goals
was to *not* block the UI thread in VS this might well cause a
problem. Also the Local final action in `Parallel.ForEach()` is run
for every task, not when all have completed. As a result the
`Complete()` method was being called many times. Luckily this did not
cause a problem because `Parallel.ForEach()` was blocking and
`base.Execute` (which does the waiting in `AsyncTask`) wasn't running.

Taking notes from the `Aot.cs` task, we should kick off the
processing in a thread of its own. This will ensure that we
(a) don't block the UI thread and (b) make sure tasks are run off the
UI Thread and (c) makes sure `Complete()` is called appropriately.

Finally, we [should be using][0] `TaskScheduler.Default` not
`TaskScheduler.Current` on the new thread.

[0]: https://github.com/Microsoft/vs-threading/blob/master/doc/threading_rules.md