[Xamarin.Android.Build.Tasks] VS & NetStandard refs (#1428)
On a Visual Studio project which has an Android App referencing a
NetStandard library, making changes on the NetStandard library will
not force a build of either project.

The cause of the problem is that the `<UsingTask/>` import for the
`<GetDependsOnNETStandard/>` task is *conditionally defined*, and the
`<UsingTask/>` is *always* evaluated:

	<!-- Bad! -->
	<!-- File: Example.targets -->
	<UsingTask
	    Condition="@CONDITION@"
	    TaskName="GetDependsOnNETStandard"
	    AssemblyFile="..." />	

If `Example.targets` is imported multiple times, and `@CONDITION@`
is True on one import but False later, then the last one "wins".
The result is that the task can be *redefined* to garbage.

Fix this by moving the `<UsingTask/>` into
`Xamarin.Android.PCLSupport.props`, where it is *unconditional*, and
moving the original Condition into the `<Import/>` of
`Xamarin.Android.PCLSupport.props`. This allows the
`<GetDependsOnNETStandard/>` to either exist or not exist at all,
not some intermediate state where it's defined to garbage.

This can be easily tried with:

	https://github.com/Microsoft/SmartHotel360-mobile-desktop-apps

Open the Android Solution and run, then change the `Login.xaml` and
try running again. Without this change, the deployed app will not be
affected. It works correctly with this change.