[Xamarin.Android.Build.Tasks] Add Locking for processing aapt output.
We got a bug report where Aapt crashes with the following

	Unhandled Exception: System.ArgumentOutOfRangeException: capacity was less than the current size.
	Parameter name: value
	   at System.ThrowHelper.ThrowArgumentOutOfRangeException(ExceptionArgument argument, ExceptionResource resource)
	   at System.Collections.Generic.List`1.set_Capacity(Int32 value)
	   at System.Collections.Generic.List`1.EnsureCapacity(Int32 min)
	   at System.Collections.Generic.List`1.Add(T item)
	   at Xamarin.Android.Tasks.Aapt.<>c__DisplayClass134_0.<RunAapt>b__0(Object sender, DataReceivedEventArgs e)

This looks like a problem were the StdOut and StdError message handlers
are being called a the same time. As a result they try to add and
item at the same time causing this problem when the collection tries
to grow.

So lets put in some locks to ensure that only one message handler
can add an item at a time.